Implementacao chamada à busca por similaridade em Servlet..continuação...Enfrentando problemas para pegar o valor da similaridade....

+++ src/br/edu/fei/sigepapp/bancodedados/model/AtributosBuscaSimilaridade.java
@@ -11,7 +11,7 @@ package br.edu.fei.sigepapp.bancodedados.model;
 public class AtributosBuscaSimilaridade {
 
     long cd_objeto;
-    long similaridade;
+    double similaridade;
     String tp_estrutura;
     String nm_objeto;
     String contexto;
@@ -22,7 +22,7 @@ public class AtributosBuscaSimilaridade {
     public AtributosBuscaSimilaridade() {
     }
 
-    public AtributosBuscaSimilaridade(long cd_objeto, long similaridade,String tp_estrutura, String nm_objeto, String contexto, String problema, String solucao) {
+    public AtributosBuscaSimilaridade(long cd_objeto, double similaridade,String tp_estrutura, String nm_objeto, String contexto, String problema, String solucao) {
         this.cd_objeto = cd_objeto;
         this.similaridade = similaridade;
         this.tp_estrutura=tp_estrutura;
@@ -64,11 +64,11 @@ public class AtributosBuscaSimilaridade {
         this.problema = problema;
     }
 
-    public long getSimilaridade() {
+    public double getSimilaridade() {
         return similaridade;
     }
 
-    public void setSimilaridade(long similaridade) {
+    public void setSimilaridade(double similaridade) {
         this.similaridade = similaridade;
     }
